+++ treasury/src/app/app.module.ts
@@ -25,10 +25,10 @@ import { ErrorPageDefaultComponent } from './error-page-default/error-page-defau
 import { FooterComponent } from './footer/footer.component';
 import { FormAddComponent } from './form-add/form-add.component';
 import { HeaderComponent } from './header/header.component';
+import { ImprintComponent } from './imprint/imprint.component';
 import { HomeComponent } from './home/home.component';
 import { MovieListComponent } from './movie-list/movie-list.component';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
-import { PrivacyPolicyComponent } from './privacy-policy/privacy-policy.component';
 import { SettingsComponent } from './settings/settings.component';
 import { TheMovieDbService } from './themoviedb/the-movie-db.service';
 import { UserService } from './services/user.service';
@@ -58,7 +58,7 @@ export class RavenErrorHandler implements ErrorHandler {
     BaseComponent,
     ErrorPageDefaultComponent,
     FooterComponent,
-    PrivacyPolicyComponent
+    ImprintComponent
   ],
   entryComponents: [ErrorComponent],
   imports: [